refactor: format painter
